Move cloudinary functions to 'services' to have a better separation of concerns.
