+++ server/module-alias.js
@@ -7,5 +7,6 @@ moduleAlias.addAliases({
     "@middlewares": __dirname + "/middlewares",
     "@models": __dirname + "/models",
     "@routes": __dirname + "/routes",
+    "@services": __dirname + "/services",
     "@utils": __dirname + "/utils"
 });
